# Add custom ring implementation to the BatchProcessor