Part of #5063

The current queue implementation uses the built-in `container/ring` package for a ring implementation. This implementation stores its value as an `any`. Each time the batch processor stores a `Record` to this field it needs to be allocated to the heap to convert it to an `any` type.

This changes the implementation to be a paired-down ring with a `Record` as the value field. This removes the allocation and greatly improves the performance of the batch processor.

### Benchmarks

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/log
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
                       │   old.txt    │               new.txt               │
                       │    sec/op    │   sec/op     vs base                │
BatchProcessorOnEmit-8   312.3n ± 18%   118.0n ± 5%  -62.22% (p=0.000 n=10)

                       │    old.txt    │                new.txt                │
                       │      B/s      │     B/s       vs base                 │
BatchProcessorOnEmit-8   1.264Gi ± 15%   3.347Gi ± 5%  +164.69% (p=0.000 n=10)

                       │  old.txt   │              new.txt              │
                       │    B/op    │   B/op    vs base                 │
BatchProcessorOnEmit-8   416.0 ± 0%   0.0 ± 0%  -100.00% (p=0.000 n=10)

                       │  old.txt   │               new.txt               │
                       │ allocs/op  │ allocs/op   vs base                 │
BatchProcessorOnEmit-8   1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
```